Corrected hyperlink to the second Downloads button
Hi Team,

I just discovered this awesome project but on your website, the second button to Downloads was hyperlinked to "/docs/general/#getting-started", which was throwing a 404. I saw that the first button on that page was hyperlinked to "/downloads" so I tried to find this page's source. I could only find reference to that button in this page so I am opening a PR to change the hyperlink to "/downloads" instead.